+++ CHANGELOG.md
@@ -10,6 +10,16 @@ This changelog is following the recommended format by [keepachangelog](https://k
 
 ### Added
 
+### Changed
+
+### Fixed
+
+### Removed
+
+## [1.3.22] - 2025-10-03
+
+### Added
+
 - Command to clone a transform
 
 ### Changed
@@ -20,7 +30,6 @@ This changelog is following the recommended format by [keepachangelog](https://k
 
 - Issue Attribute Sync Config export in multiple files if the source have been deleted (cf. [#123](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/issues/123))
 
-### Removed
 
 ## [1.3.21] - 2025-09-24